+++ ReleaseNotes.md
@@ -2,6 +2,10 @@
 
 # Release 2.2.2
 
+## Features
+
+* TGR-912: Healthchecks for externalUrl-Servers now honor the configured forwardProxyInfo of the localTigerProxy
+
 ## Bugfixes
 * TGR-1139/TGR-1140: Fixed various minor problems with recursive descent and attributes in RbelPath-expressions
 * TGR-1038: Changed assertThatNo()-exceptions to assertThatNo.isThrownBy so the tests are not ignored